merge: resolve conflicts with main (program moved to cli.ts; version 1.44.0)
