+++ bin/cli.js
@@ -1,3 +1,5 @@
 #!/usr/bin/env node
 
-import '../dist/index.js';
+import { program } from '../dist/index.js';
+
+program.parse();
