+++ src/commands/login.ts
@@ -277,7 +277,7 @@ export async function loginCommand(): Promise<void> {
   console.log(pc.cyan('  ┌─────────────────────────────────────────────────────────────'));
   console.log(pc.cyan('  │'));
   console.log(pc.cyan('  │  The One CLI (`one`) is now installed on this machine. It'));
-  console.log(pc.cyan('  │  lets you interact with 250+ third-party platforms (Gmail,'));
+  console.log(pc.cyan('  │  lets you interact with 400+ third-party platforms (Gmail,'));
   console.log(pc.cyan('  │  Slack, Shopify, HubSpot, Stripe, GitHub, etc.) through'));
   console.log(pc.cyan('  │  their APIs. Whenever you need to interact with any'));
   console.log(pc.cyan('  │  third-party platform or external service, use the `one`'));
