+++ src/commands/onboard.ts
@@ -51,7 +51,7 @@ function buildStep1(): { step: number; totalSteps: number; message: string; next
     nextCommand: 'one onboard --step 2',
     message: `# One CLI — Agent Onboarding
 
-You now have access to the One CLI. It connects you to 250+ platforms
+You now have access to the One CLI. It connects you to 400+ platforms
 (Gmail, Slack, Stripe, Notion, etc.) — you'll learn the details shortly.
 
 IMPORTANT: Your only task right now is to send the message below to the
@@ -61,7 +61,7 @@ Do not offer demos yet. Just send this message and wait for their reply:
 ---
 
 I just got access to the One CLI — it connects me to your platforms like
-Gmail, Slack, Stripe, and 250+ others so I can work with them directly.
+Gmail, Slack, Stripe, and 400+ others so I can work with them directly.
 I'll show you what I can do in a moment.
 
 First, one quick setup step: right now only this conversation knows about
@@ -96,7 +96,7 @@ If they declined, skip this section.
 ## One CLI — Platform Integrations
 
 The One CLI (\`one\`) is installed and configured on this machine. It is a
-command-line tool that gives AI agents the ability to interact with 250+
+command-line tool that gives AI agents the ability to interact with 400+
 third-party platforms (Gmail, Slack, Shopify, HubSpot, Stripe, GitHub,
 Notion, Salesforce, etc.) through a single, unified interface.
 
@@ -201,7 +201,7 @@ function buildStep3(connections: Connection[]): { step: number; totalSteps: numb
 The user can connect any of these popular platforms:
 Gmail, Google Calendar, Slack, Notion, Shopify, Stripe, HubSpot, GitHub,
 Salesforce, QuickBooks, Asana, Jira, Linear, Intercom, Zendesk, Twilio,
-and 250+ more. Run \`one platforms\` for the full list.`);
+and 400+ more. Run \`one platforms\` for the full list.`);
 
   sections.push('## Onboarding complete!\n\nYou\'re all set. Use `one --agent guide` any time you need the full reference.');
 
@@ -236,7 +236,7 @@ platform before you can start using actions. Suggest they run:
   one add slack
   one add <any-platform>
 
-Run \`one platforms\` to see all 250+ available platforms.`;
+Run \`one platforms\` to see all 400+ available platforms.`;
   }
 
   const header = `## Current State
@@ -259,7 +259,7 @@ function buildDemoActions(connections: Connection[]): string {
   lines.push('Try these to prove it works:');
 
   const connectedPlatforms = connections.map(c => c.platform.toLowerCase());
-  const popularPlatforms = ['gmail', 'google-calendar', 'slack', 'shopify', 'hub-spot', 'github'];
+  const popularPlatforms = ['gmail', 'google-calendar', 'slack', 'shopify', 'hubspot', 'github'];
 
   const platformsToShow = [
     ...connectedPlatforms.filter(p => PLATFORM_DEMO_ACTIONS[p]),
