+++ src/lib/platform-meta.ts
@@ -5,7 +5,7 @@ export const PLATFORM_DEMO_ACTIONS: Record<string, { query: string; description:
   'google-calendar': { query: 'list events', description: "Check today's schedule" },
   slack: { query: 'list channels', description: 'List Slack channels' },
   shopify: { query: 'list orders', description: 'List recent orders' },
-  'hub-spot': { query: 'list contacts', description: 'Search CRM contacts' },
+  hubspot: { query: 'list contacts', description: 'Search CRM contacts' },
   github: { query: 'list repositories', description: 'List repos' },
   notion: { query: 'list pages', description: 'List Notion pages' },
   stripe: { query: 'list payments', description: 'List recent payments' },
@@ -23,10 +23,10 @@ export const PLATFORM_DEMO_ACTIONS: Record<string, { query: string; description:
 const WORKFLOW_EXAMPLES: Record<string, string> = {
   'gmail+google-calendar': 'Check my calendar for today and draft a summary email',
   'gmail+shopify': 'Find unfulfilled orders and email each customer an update',
-  'hub-spot+slack': 'Find deals closing this week and post a summary to Slack',
+  'hubspot+slack': 'Find deals closing this week and post a summary to Slack',
   'github+slack': 'List open PRs and post a review reminder to #engineering',
   'gmail+stripe': 'Find failed payments this week and send retry reminder emails',
-  'gmail+hub-spot': 'Find new CRM contacts and send them a welcome email',
+  'gmail+hubspot': 'Find new CRM contacts and send them a welcome email',
   'notion+slack': 'Summarize recent Notion updates and post to a Slack channel',
   'github+jira': 'Link recent commits to Jira issues and update their status',
   'google-calendar+slack': "Post today's meeting schedule to a Slack channel",
